# Switch to @swc/jest to speed up tests Fixes #1679