--- a/package.json
+++ b/package.json
@@ -29,6 +29,7 @@
     "@datastructures-js/priority-queue": "^6.3.3",
     "@eslint/compat": "^1.2.7",
     "@eslint/js": "^9.21.0",
+    "@swc/jest": "^0.2.39",
     "@types/benchmark": "^2.1.5",
     "@types/chai": "^4.3.17",
     "@types/d3": "^7.4.3",