--- a/package.json
+++ b/package.json
@@ -85,7 +85,6 @@
     "sinon-chai": "^4.0.0",
     "style-loader": "^4.0.0",
     "tailwindcss": "^3.4.17",
-    "ts-jest": "^29.2.4",
     "ts-loader": "^9.5.2",
     "tsconfig-paths": "^4.2.0",
     "tsx": "^4.17.0",