@CLAassistant 2025-08-02
[![CLA assistant check](https://cla-assistant.io/pull/badge/not_signed)](https://cla-assistant.io/openfrontio/OpenFrontIO?pullRequest=1680) <br/>Thank you for your submission! We really appreciate it. Like many open source projects, we ask that you all sign our [Contributor License Agreement](https://cla-assistant.io/openfrontio/OpenFrontIO?pullRequest=1680) before we can accept your contribution.<br/>**1** out of **2** committers have signed the CLA.<br/><br/>:white_check_mark: TylerHanavan<br/>:x: scottanderson<br/><sub>You have signed the CLA already but the status is still pending? Let us [recheck](https://cla-assistant.io/check/openfrontio/OpenFrontIO?pullRequest=1680) it.</sub>